ai_course_bot/ai_chatbot_backend/requirements.txt
fastapi==0.111.0
FlagEmbedding==1.2.10
numpy==1.26.4
openai==1.31.0
pydantic==2.7.3
python-dotenv==1.0.1
sqlite_vss==0.1.2
starlette==0.37.2
torch==2.3.0
transformers==4.40.1
uvicorn==0.30.1

cachetools==5.5.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
requests-oauthlib==2.0.0
rsa==4.9

pydantic-settings==2.8.1
python-dotenv>=0.21.0